requirements.txt
django
submission
django-cors-headers
djangorestframework
psycopg2